@@ -13,7 +13,8 @@
 # limitations under the License.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, List, Sequence, Tuple, Union
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Union
 
 from typing_extensions import TypeAlias
 
@@ -34,10 +35,10 @@ _StaticShapeLike: TypeAlias = Union[
 ShapeLike: TypeAlias = Union[_DynamicShapeLike, _StaticShapeLike]
 
 # for size parameters, eg, kernel_size, stride ...
-Size1: TypeAlias = Union[int, Tuple[int], List[int]]
-Size2: TypeAlias = Union[int, Tuple[int, int], List[int]]
-Size3: TypeAlias = Union[int, Tuple[int, int, int], List[int]]
-Size4: TypeAlias = Union[int, Tuple[int, int, int, int], List[int]]
-Size5: TypeAlias = Union[int, Tuple[int, int, int, int, int], List[int]]
-Size6: TypeAlias = Union[int, Tuple[int, int, int, int, int, int], List[int]]
-SizeN: TypeAlias = Union[int, Tuple[int, ...], List[int]]
+Size1: TypeAlias = Union[int, tuple[int], list[int]]
+Size2: TypeAlias = Union[int, tuple[int, int], list[int]]
+Size3: TypeAlias = Union[int, tuple[int, int, int], list[int]]
+Size4: TypeAlias = Union[int, tuple[int, int, int, int], list[int]]
+Size5: TypeAlias = Union[int, tuple[int, int, int, int, int], list[int]]
+Size6: TypeAlias = Union[int, tuple[int, int, int, int, int, int], list[int]]
+SizeN: TypeAlias = Union[int, tuple[int, ...], list[int]]
