@@ -20,12 +20,8 @@ from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
-    ContextManager,
-    List,
     Literal,
     Protocol,
-    Set,
-    Tuple,
     TypeVar,
     Union,
     overload,
@@ -46,7 +42,8 @@ from paddle.static.amp.decorator import OptimizerWithMixedPrecision
 from .amp_lists import black_list, white_list
 
 if TYPE_CHECKING:
-    from typing import Generator
+    from collections.abc import Generator
+    from contextlib import AbstractContextManager
 
     from typing_extensions import TypeAlias, TypeGuard
 
@@ -57,7 +54,7 @@ if TYPE_CHECKING:
     from paddle.static import Operator, Program
 
     _AmpLevelLiteral = Literal["O0", "OD", "O1", "O2"]
-    _CustomList: TypeAlias = Union[List[str], Tuple[str, ...], Set[str]]
+    _CustomList: TypeAlias = Union[list[str], tuple[str, ...], set[str]]
 
     class _OptimizerLike(Protocol):
         def minimize(
@@ -75,8 +72,8 @@ if TYPE_CHECKING:
         def clear_grad(self, set_to_zero: bool) -> None: ...
 
 
-_ModelsT = TypeVar("_ModelsT", "Layer", List["Layer"])
-_OptimizersT = TypeVar("_OptimizersT", "_OptimizerLike", List["_OptimizerLike"])
+_ModelsT = TypeVar("_ModelsT", "Layer", list["Layer"])
+_OptimizersT = TypeVar("_OptimizersT", "_OptimizerLike", list["_OptimizerLike"])
 
 
 AMP_RELATED_FLAGS = [
@@ -1066,7 +1063,7 @@ def auto_cast(
     level: _AmpLevelLiteral = 'O1',
     dtype: _DTypeLiteral = 'float16',
     use_promote: bool = True,
-) -> ContextManager:
+) -> AbstractContextManager:
     """
     Create a context which enables auto-mixed-precision(AMP) of operators executed in dynamic graph mode.
     If enabled, the input data type (float32, float16 or bfloat16) of each operator is decided
