@@ -23,7 +23,6 @@ import warnings
 from typing import (
     TYPE_CHECKING,
     Any,
-    List,
     Literal,
     Union,
     overload,
@@ -70,8 +69,8 @@ if TYPE_CHECKING:
     _InputBatch: TypeAlias = Union[
         Tensor,
         npt.NDArray[Any],
-        List[Tensor],
-        List[npt.NDArray[Any]],
+        list[Tensor],
+        list[npt.NDArray[Any]],
     ]
 
 
