@@ -18,14 +18,14 @@ import copy
 import inspect
 import operator
 from functools import cached_property, reduce
-from typing import TYPE_CHECKING, Any, Callable, Dict, Tuple, TypeVar
+from typing import TYPE_CHECKING, Any, Callable, TypeVar
 
 from ...utils import InnerError, NameGenerator, hashable
 
 if TYPE_CHECKING:
     T = TypeVar("T")
-    Args = Tuple[T, ...]
-    Kwargs = Dict[str, T]
+    Args = tuple[T, ...]
+    Kwargs = dict[str, T]
 
 
 def format_type(type_: type[Any] | tuple[type[Any], ...]) -> str:
